requirements.txt
azure-ai-vision-imageanalysis==1.0.0
python-dotenv
pillow
